Move database to its own folder, as we need to split it off so we can add the one for the Shotwell DB.
