+++ src/myarchive/db/tag_db/ljtables.py
@@ -3,10 +3,10 @@ from sqlalchemy import (
 from sqlalchemy.orm import backref, relationship
 from sqlalchemy.orm.exc import NoResultFound
 
-from myarchive.db.tables.association_tables import (
+from myarchive.db.tag_db.association_tables import (
     at_ljcomment_tag, at_ljentry_tag)
-from myarchive.db.tables.base import Base
-from myarchive.db.tables.tag import Tag
+from myarchive.db.tag_db.base import Base
+from myarchive.db.tag_db.tag import Tag
 
 
 class CircularDependencyError(Exception):
